+++ src/KDiskCache.cpp
@@ -32,7 +32,7 @@ index_scan_state_t index_scan_state;
 static int load_count = 0;
 static INT64 recreate_start_time = 0;
 using namespace std;
-bool obj_can_disk_cache(KHttpRequest *rq, KHttpObject *obj)
+bool obj_can_disk_cache(KHttpRequest* rq, KHttpObject* obj)
 {
 	if (KBIT_TEST(rq->filter_flags, RF_NO_DISK_CACHE)) {
 		return false;
@@ -42,129 +42,129 @@ bool obj_can_disk_cache(KHttpRequest *rq, KHttpObject *obj)
 	}
 	return cache.IsDiskCacheReady();
 }
-bool skipString(char **hot,int &hotlen)
+bool skipString(char** hot, int& hotlen)
 {
-	if (hotlen<=(int)sizeof(int)) {
+	if (hotlen <= (int)sizeof(int)) {
 		return false;
 	}
 	int len;
-	kgl_memcpy(&len,*hot,sizeof(int));
-	(*hot)+=sizeof(int);
-	hotlen-=sizeof(int);
-	if (hotlen<=len) {
+	kgl_memcpy(&len, *hot, sizeof(int));
+	(*hot) += sizeof(int);
+	hotlen -= sizeof(int);
+	if (hotlen <= len) {
 		return false;
 	}
 	(*hot) += len;
 	hotlen -= len;
 	return true;
 }
-bool skipString(KFile *file)
+bool skipString(KFile* file)
 {
 	int len;
-	if(file->read((char *)&len,sizeof(len))!=sizeof(len)){
+	if (file->read((char*)&len, sizeof(len)) != sizeof(len)) {
 		return false;
 	}
-	return file->seek(len,seekCur);
+	return file->seek(len, seekCur);
 }
-char *readString(char **hot,int &hotlen,int &len)
+char* readString(char** hot, int& hotlen, int& len)
 {
-	if (hotlen<(int)sizeof(int)) {
+	if (hotlen < (int)sizeof(int)) {
 		len = -1;
 		return NULL;
 	}
-	kgl_memcpy(&len,*hot,sizeof(int));
-	hotlen-=sizeof(int);
+	kgl_memcpy(&len, *hot, sizeof(int));
+	hotlen -= sizeof(int);
 	(*hot) += sizeof(int);
-	if(len<0 || len>1000000){
-		klog(KLOG_ERR,"string len[%d] is too big\n",len);
+	if (len < 0 || len>1000000) {
+		klog(KLOG_ERR, "string len[%d] is too big\n", len);
 		len = -1;
 		return NULL;
 	}
-	if (hotlen<len) {
+	if (hotlen < len) {
 		len = -1;
 		return NULL;
 	}
-	char *buf = (char *)xmalloc(len+1);
-	buf[len]='\0';
-	if (len>0) {
-		kgl_memcpy(buf,*hot,len);
+	char* buf = (char*)xmalloc(len + 1);
+	buf[len] = '\0';
+	if (len > 0) {
+		kgl_memcpy(buf, *hot, len);
 		hotlen -= len;
 		(*hot) += len;
 	}
 	return buf;
 }
-char *readString(KFile *file,int &len)
+char* readString(KFile* file, int& len)
 {
-	if(file->read((char *)&len,sizeof(len))!=sizeof(len)){
+	if (file->read((char*)&len, sizeof(len)) != sizeof(len)) {
 		len = -1;
 		return NULL;
 	}
-	if(len<0 || len>1000000){
+	if (len < 0 || len>1000000) {
 		len = -1;
-		klog(KLOG_ERR,"string len[%d] is too big\n",len);
+		klog(KLOG_ERR, "string len[%d] is too big\n", len);
 		return NULL;
 	}
-	char *buf = (char *)xmalloc(len+1);
-	buf[len]='\0';
-	if (len>0 && (int)file->read(buf,len)!=len) {
+	char* buf = (char*)xmalloc(len + 1);
+	buf[len] = '\0';
+	if (len > 0 && (int)file->read(buf, len) != len) {
 		xfree(buf);
 		len = -1;
 		return NULL;
 	}
 	return buf;
 }
-int write_string(char *hot, const char *str, int len)
+int write_string(char* hot, const char* str, int len)
 {
-	kgl_memcpy(hot, (char *)&len, sizeof(int));
+	kgl_memcpy(hot, (char*)&len, sizeof(int));
 	if (len > 0) {
 		kgl_memcpy(hot + sizeof(int), str, len);
 	}
 	return len + sizeof(int);
 }
-int writeString(KBufferFile *file,const char *str,int len)
+int writeString(KBufferFile* file, const char* str, int len)
 {
 	if (str) {
-		if (len==0) {
-			len = strlen(str);
+		if (len == 0) {
+			len = (int)strlen(str);
 		}
 	}
-	int ret = file->write((char *)&len,sizeof(len));
-	if (len>0) {
-		ret += file->write(str,len);
+	int ret = file->write((char*)&len, sizeof(len));
+	if (len > 0) {
+		ret += file->write(str, len);
 	}
 	return ret;
 }
-bool read_obj_head(KHttpObjectBody *data,char **hot,int &hotlen)
+bool read_obj_head(KHttpObjectBody* data, char** hot, int& hotlen)
 {
-	assert(data->headers==NULL);
-	KHttpHeader *last = NULL;
+	assert(data->headers == NULL);
+	KHttpHeader* last = NULL;
 	for (;;) {
-		int attr_len,val_len;
+		int attr_len, val_len;
 		//printf("before attr hotlen=[%d]\n",hotlen);
-		char *attr = readString(hot,hotlen,attr_len);
+		char* attr = readString(hot, hotlen, attr_len);
 		//printf("after attr before val hotlen=[%d]\n",hotlen);
-		if (attr_len==-1) {
+		if (attr_len == -1) {
 			return false;
 		}
-		if (attr==NULL) {
+		if (attr == NULL) {
 			return true;
 		}
-		if (*attr=='\0') {
+		if (*attr == '\0') {
 			free(attr);
 			return true;
 		}
 		//printf("attr=[%s]\n",attr);
-		char *val = readString(hot,hotlen,val_len);
+		char* val = readString(hot, hotlen, val_len);
 		//printf("after val hotlen=[%d]\n",hotlen);
-		if(val_len==-1){
+		if (val_len == -1) {
 			xfree(attr);
 			return false;
 		}
 		//printf("val=[%s]\n",val);
-		KHttpHeader *header = (KHttpHeader *)xmalloc(sizeof(KHttpHeader));
-		if(header==NULL){
+		KHttpHeader* header = (KHttpHeader*)xmalloc(sizeof(KHttpHeader));
+		if (header == NULL) {
 			xfree(attr);
-			if(val){
+			if (val) {
 				xfree(val);
 			}
 			return false;
@@ -174,40 +174,40 @@ bool read_obj_head(KHttpObjectBody *data,char **hot,int &hotlen)
 		header->next = NULL;
 		header->attr_len = attr_len;
 		header->val_len = val_len;
-		if(last==NULL){
-			data->headers = header;		
-		}else{
+		if (last == NULL) {
+			data->headers = header;
+		} else {
 			last->next = header;
 		}
 		last = header;
 	}
 }
-bool read_obj_head(KHttpObjectBody *data,KFile *fp)
+bool read_obj_head(KHttpObjectBody* data, KFile* fp)
 {
-	assert(data->headers==NULL);
-	KHttpHeader *hot = NULL;
+	assert(data->headers == NULL);
+	KHttpHeader* hot = NULL;
 	for (;;) {
-		int attr_len,val_len;
-		char *attr = readString(fp,attr_len);
+		int attr_len, val_len;
+		char* attr = readString(fp, attr_len);
 		if (attr_len == -1) {
 			return false;
 		}
-		if (attr==NULL) {
+		if (attr == NULL) {
 			return true;
 		}
-		if (*attr=='\0') {
+		if (*attr == '\0') {
 			free(attr);
 			return true;
 		}
-		char *val = readString(fp,val_len);
-		if(val_len==-1){
+		char* val = readString(fp, val_len);
+		if (val_len == -1) {
 			xfree(attr);
 			return false;
 		}
-		KHttpHeader *header = (KHttpHeader *)xmalloc(sizeof(KHttpHeader));
-		if(header==NULL){
+		KHttpHeader* header = (KHttpHeader*)xmalloc(sizeof(KHttpHeader));
+		if (header == NULL) {
 			xfree(attr);
-			if(val){
+			if (val) {
 				xfree(val);
 			}
 			return false;
@@ -217,31 +217,31 @@ bool read_obj_head(KHttpObjectBody *data,KFile *fp)
 		header->attr_len = attr_len;
 		header->val_len = val_len;
 		header->next = NULL;
-		if(hot==NULL){
-			data->headers = header;		
-		}else{
+		if (hot == NULL) {
+			data->headers = header;
+		} else {
 			hot->next = header;
 		}
 		hot = header;
 	}
 }
-char *getCacheIndexFile()
+char* getCacheIndexFile()
 {
 	KStringBuf s;
 	if (*conf.disk_cache_dir) {
 		s << conf.disk_cache_dir;
 	} else {
-		s << conf.path << "cache" << PATH_SPLIT_CHAR ;
+		s << conf.path << "cache" << PATH_SPLIT_CHAR;
 	}
 	s << "index";
 	return s.stealString();
 }
-void get_index_scan_state_filename(KStringBuf &s)
+void get_index_scan_state_filename(KStringBuf& s)
 {
 	if (*conf.disk_cache_dir) {
 		s << conf.disk_cache_dir;
 	} else {
-		s << conf.path << "cache" << PATH_SPLIT_CHAR ;
+		s << conf.path << "cache" << PATH_SPLIT_CHAR;
 	}
 	s << "index.scan";
 }
@@ -250,11 +250,11 @@ bool save_index_scan_state()
 	KStringBuf s;
 	get_index_scan_state_filename(s);
 	KFile fp;
-	if (!fp.open(s.getString(),fileWrite)) {
+	if (!fp.open(s.getString(), fileWrite)) {
 		return false;
 	}
 	bool result = true;
-	if (sizeof(index_scan_state_t) != fp.write((char *)&index_scan_state,sizeof(index_scan_state_t))) {
+	if (sizeof(index_scan_state_t) != fp.write((char*)&index_scan_state, sizeof(index_scan_state_t))) {
 		result = false;
 	}
 	fp.close();
@@ -265,11 +265,11 @@ bool load_index_scan_state()
 	KStringBuf s;
 	get_index_scan_state_filename(s);
 	KFile fp;
-	if (!fp.open(s.getString(),fileRead)) {
+	if (!fp.open(s.getString(), fileRead)) {
 		return false;
 	}
 	bool result = true;
-	if (sizeof(index_scan_state_t) != fp.read((char *)&index_scan_state,sizeof(index_scan_state_t))) {
+	if (sizeof(index_scan_state_t) != fp.read((char*)&index_scan_state, sizeof(index_scan_state_t))) {
 		result = false;
 	}
 	fp.close();
@@ -288,7 +288,7 @@ bool saveCacheIndex()
 	cache.syncDisk();
 #ifdef ENABLE_DB_DISK_INDEX
 	if (dci) {
-		dci->start(ci_close,NULL);
+		dci->start(ci_close, NULL);
 		while (!dci->allWorkedDone()) {
 			my_msleep(100);
 		}
@@ -297,9 +297,9 @@ bool saveCacheIndex()
 #endif
 	return false;
 }
-cor_result create_http_object2(KHttpObject *obj, char *url, uint32_t flag_encoding, const char *verified_filename)
+cor_result create_http_object2(KHttpObject* obj, char* url, uint32_t flag_encoding, const char* verified_filename)
 {
-	char *vary = strchr(url, '\n');
+	char* vary = strchr(url, '\n');
 	if (vary != NULL) {
 		*vary = '\0';
 		vary++;
@@ -319,7 +319,7 @@ cor_result create_http_object2(KHttpObject *obj, char *url, uint32_t flag_encodi
 	obj->uk.url->port = m_url.port;
 	obj->uk.url->flag_encoding = flag_encoding;
 	if (vary) {
-		char *vary_val = strchr(vary, '\n');
+		char* vary_val = strchr(vary, '\n');
 		if (vary_val) {
 			*vary_val = '\0';
 			vary_val++;
@@ -338,8 +338,8 @@ cor_result create_http_object2(KHttpObject *obj, char *url, uint32_t flag_encodi
 	}
 	if (KBIT_TEST(obj->index.flags, FLAG_IN_DISK)) {
 		if (obj->index.head_size != kgl_align(obj->index.head_size, kgl_aio_align_size)) {
-			char *url = obj->uk.url->getUrl();
-			char *filename = obj->getFileName();
+			char* url = obj->uk.url->getUrl();
+			char* filename = obj->getFileName();
 			klog(KLOG_ERR, "disk cache file head_size=[%d] is not align by size=[%d], url=[%s] file=[%s] now dead it.\n", obj->index.head_size, kgl_aio_align_size, url, filename);
 			free(filename);
 			free(url);
@@ -347,8 +347,8 @@ cor_result create_http_object2(KHttpObject *obj, char *url, uint32_t flag_encodi
 		}
 #ifndef ENABLE_BIG_OBJECT_206
 		if (KBIT_TEST(obj->index.flags, FLAG_BIG_OBJECT_PROGRESS)) {
-			char *url = obj->uk.url->getUrl();
-			char *filename = obj->getFileName();
+			char* url = obj->uk.url->getUrl();
+			char* filename = obj->getFileName();
 			klog(KLOG_ERR, "disk cache file head_size=[%d] is part file that is not support by now size=[%d], url=[%s] file=[%s]. now dead it.\n", obj->index.head_size, kgl_aio_align_size, url, filename);
 			free(filename);
 			free(url);
@@ -362,48 +362,48 @@ cor_result create_http_object2(KHttpObject *obj, char *url, uint32_t flag_encodi
 	KBIT_CLR(obj->index.flags, FLAG_IN_DISK);
 	return cor_failed;
 }
-cor_result create_http_object(KHttpObject *obj,const char *url, uint32_t flag_encoding,const char *verified_filename)
+cor_result create_http_object(KHttpObject* obj, const char* url, uint32_t flag_encoding, const char* verified_filename)
 {
 	//printf("url=[%s]\n", url);
-	char *buf = strdup(url);
+	char* buf = strdup(url);
 	cor_result ret = create_http_object2(obj, buf, flag_encoding, verified_filename);
 	xfree(buf);
 	return ret;
 }
-cor_result create_http_object(KFile *fp,KHttpObject *obj,uint32_t url_flag_encoding,const char *verified_filename=NULL)
+cor_result create_http_object(KFile* fp, KHttpObject* obj, uint32_t url_flag_encoding, const char* verified_filename = NULL)
 {
 	int len;
-	char *url = readString(fp,len);
-	if(url==NULL){
-		fprintf(stderr,"read url is NULL\n");
+	char* url = readString(fp, len);
+	if (url == NULL) {
+		fprintf(stderr, "read url is NULL\n");
 		return cor_failed;
 	}
-	cor_result ret = create_http_object2(obj,url, url_flag_encoding,verified_filename);
+	cor_result ret = create_http_object2(obj, url, url_flag_encoding, verified_filename);
 	free(url);
 	return ret;
 }
-int create_file_index(const char *file,void *param)
+int create_file_index(const char* file, void* param)
 {
 	KStringBuf s;
 	cor_result result = cor_failed;
-	KHttpObject *obj;
-	s << (char *)param << PATH_SPLIT_CHAR << file;
+	KHttpObject* obj;
+	s << (char*)param << PATH_SPLIT_CHAR << file;
 	unsigned f1 = 0;
 	unsigned f2 = 0;
-	char *file_name = s.getString();
+	char* file_name = s.getString();
 	KFile fp;
-	if (!fp.open(s.getString(),fileRead)) {
-		fprintf(stderr,"cann't open file[%s]\n",s.getString());
+	if (!fp.open(s.getString(), fileRead)) {
+		fprintf(stderr, "cann't open file[%s]\n", s.getString());
 		return 0;
 	}
-	if (recreate_start_time>0) {
+	if (recreate_start_time > 0) {
 		INT64 t = fp.getCreateTime();
-		if (t>recreate_start_time) {
-			klog(KLOG_DEBUG,"file [%s] is new file t=%d\n",file_name,(int)(t-recreate_start_time));
+		if (t > recreate_start_time) {
+			klog(KLOG_DEBUG, "file [%s] is new file t=%d\n", file_name, (int)(t - recreate_start_time));
 			return 0;
 		}
 	}
-	if (strstr(file,".part")) {
+	if (strstr(file, ".part")) {
 #ifdef ENABLE_BIG_OBJECT_206
 		std::string obj_file = file;
 		obj_file = obj_file.substr(0, obj_file.size() - 5);
@@ -412,12 +412,12 @@ int create_file_index(const char *file,void *param)
 #endif
 		goto failed;
 	}
-	if (sscanf(file, "%x_%x", &f1, &f2)!=2) {
+	if (sscanf(file, "%x_%x", &f1, &f2) != 2) {
 		goto failed;
 	}
 	KHttpObjectFileHeader header;
-	if(fp.read((char *)&header,sizeof(KHttpObjectFileHeader))!=sizeof(KHttpObjectFileHeader)){
-		fprintf(stderr,"cann't read head size [%s]\n",file_name);
+	if (fp.read((char*)&header, sizeof(KHttpObjectFileHeader)) != sizeof(KHttpObjectFileHeader)) {
+		fprintf(stderr, "cann't read head size [%s]\n", file_name);
 		goto failed;
 	}
 	if (!is_valide_dc_head_size(header.index.head_size)) {
@@ -435,7 +435,7 @@ int create_file_index(const char *file,void *param)
 	}
 	*/
 	obj = new KHttpObject;
-	kgl_memcpy(&obj->index,&header.index,sizeof(obj->index));
+	kgl_memcpy(&obj->index, &header.index, sizeof(obj->index));
 	obj->dk.filename1 = f1;
 	obj->dk.filename2 = f2;
 	//{{ent
@@ -444,11 +444,11 @@ int create_file_index(const char *file,void *param)
 		partobjs[file] = true;
 	}
 #endif//}}
-	result = create_http_object(&fp,obj,header.url_flag_encoding,file_name);
-	if (result==cor_success) {
+	result = create_http_object(&fp, obj, header.url_flag_encoding, file_name);
+	if (result == cor_success) {
 #ifdef ENABLE_DB_DISK_INDEX
 		if (dci) {
-			dci->start(ci_add,obj);
+			dci->start(ci_add, obj);
 		}
 #endif
 		load_count++;
@@ -456,13 +456,13 @@ int create_file_index(const char *file,void *param)
 	obj->release();
 failed:
 	fp.close();
-	if (result==cor_failed) {
-		klog(KLOG_NOTICE,"create http object failed,remove file[%s]\n",file_name);
-		unlink(file_name);		
+	if (result == cor_failed) {
+		klog(KLOG_NOTICE, "create http object failed,remove file[%s]\n", file_name);
+		unlink(file_name);
 	}
 	return 0;
 }
-void clean_disk_orphan_files(const char *cache_dir)
+void clean_disk_orphan_files(const char* cache_dir)
 {
 	//{{ent
 #ifdef ENABLE_BIG_OBJECT_206
@@ -481,22 +481,22 @@ void clean_disk_orphan_files(const char *cache_dir)
 	partobjs.clear();
 #endif//}}
 }
-void recreate_index_dir(const char *cache_dir)
+void recreate_index_dir(const char* cache_dir)
 {
-	klog(KLOG_NOTICE,"scan cache dir [%s]\n",cache_dir);
+	klog(KLOG_NOTICE, "scan cache dir [%s]\n", cache_dir);
 	//{{ent
 #ifdef ENABLE_BIG_OBJECT_206
 	partfiles.clear();
 	partobjs.clear();
 #endif//}}
-	list_dir(cache_dir,create_file_index,(void *)cache_dir);
+	list_dir(cache_dir, create_file_index, (void*)cache_dir);
 	clean_disk_orphan_files(cache_dir);
 }
-bool recreate_index(const char *path,int &first_dir_index,int &second_dir_index,KTimeMatch *tm=NULL)
+bool recreate_index(const char* path, int& first_dir_index, int& second_dir_index, KTimeMatch* tm = NULL)
 {
 	KStringBuf s;
-	for (;first_dir_index<=CACHE_DIR_MASK1;first_dir_index++) {
-		for (;second_dir_index<=CACHE_DIR_MASK2;second_dir_index++) {
+	for (; first_dir_index <= CACHE_DIR_MASK1; first_dir_index++) {
+		for (; second_dir_index <= CACHE_DIR_MASK2; second_dir_index++) {
 			if (tm && !tm->checkTime(time(NULL))) {
 				return false;
 			}
@@ -519,7 +519,7 @@ void recreate_index(time_t start_time)
 	}
 	recreate_start_time = start_time;
 	index_progress = true;
-	klog(KLOG_ERR,"now recreate the index file,It may be use more time.Please wait...\n");
+	klog(KLOG_ERR, "now recreate the index file,It may be use more time.Please wait...\n");
 	string path;
 	if (*conf.disk_cache_dir) {
 		path = conf.disk_cache_dir;
@@ -529,11 +529,11 @@ void recreate_index(time_t start_time)
 		path += PATH_SPLIT_CHAR;
 	}
 	KStringBuf s;
-	load_count=0;
-	int i=0;
-	int j=0;
-	recreate_index(path.c_str(),i,j,NULL);
-	klog(KLOG_ERR,"create index done. total find %d object.\n",load_count);
+	load_count = 0;
+	int i = 0;
+	int j = 0;
+	recreate_index(path.c_str(), i, j, NULL);
+	klog(KLOG_ERR, "create index done. total find %d object.\n", load_count);
 	index_progress = false;
 }
 void init_disk_cache(bool firstTime)
@@ -543,10 +543,10 @@ void init_disk_cache(bool firstTime)
 	if (dci) {
 		return;
 	}
-	memset(&index_scan_state,0,sizeof(index_scan_state));
+	memset(&index_scan_state, 0, sizeof(index_scan_state));
 	load_index_scan_state();
 	dci = new KSqliteDiskCacheIndex;
-	char *file_name = getCacheIndexFile();
+	char* file_name = getCacheIndexFile();
 	KStringBuf sqliteIndex;
 	//remove old version sqt
 	bool remove_old_index = false;
@@ -566,12 +566,12 @@ void init_disk_cache(bool firstTime)
 	sqliteIndex << file_name << ".sqt" << CACHE_DISK_VERSION;
 	free(file_name);
 	KFile fp;
-	if (fp.open(sqliteIndex.getString(),fileRead)) {
+	if (fp.open(sqliteIndex.getString(), fileRead)) {
 		fp.close();
 		if (dci->open(sqliteIndex.getString())) {
-			dci->start(ci_load,NULL);
+			dci->start(ci_load, NULL);
 		} else {
-			klog(KLOG_ERR,"recreate the disk cache index database\n");
+			klog(KLOG_ERR, "recreate the disk cache index database\n");
 			dci->close();
 			unlink(sqliteIndex.getString());
 			dci->create(sqliteIndex.getString());
@@ -584,12 +584,12 @@ void init_disk_cache(bool firstTime)
 		}
 	}
 #else
-	memset(&index_scan_state,0,sizeof(index_scan_state));
+	memset(&index_scan_state, 0, sizeof(index_scan_state));
 	load_index_scan_state();
-	m_thread.start(NULL,load_cache_index);
+	m_thread.start(NULL, load_cache_index);
 #endif
 }
-void get_disk_base_dir(KStringBuf &s)
+void get_disk_base_dir(KStringBuf& s)
 {
 	if (*conf.disk_cache_dir) {
 		s << conf.disk_cache_dir;
@@ -597,7 +597,7 @@ void get_disk_base_dir(KStringBuf &s)
 		s << conf.path << "cache" << PATH_SPLIT_CHAR;
 	}
 }
-KTHREAD_FUNCTION scan_disk_cache_thread(void *param)
+KTHREAD_FUNCTION scan_disk_cache_thread(void* param)
 {
 	string path;
 	if (*conf.disk_cache_dir) {
@@ -607,7 +607,7 @@ KTHREAD_FUNCTION scan_disk_cache_thread(void *param)
 		path += "cache";
 		path += PATH_SPLIT_CHAR;
 	}
-	if (recreate_index(path.c_str(),index_scan_state.first_index,index_scan_state.second_index,&conf.diskWorkTime)) {
+	if (recreate_index(path.c_str(), index_scan_state.first_index, index_scan_state.second_index, &conf.diskWorkTime)) {
 		index_scan_state.need_index_progress = 0;
 		index_scan_state.last_scan_time = time(NULL);
 		save_index_scan_state();
@@ -639,7 +639,7 @@ void rescan_disk_cache()
 	index_scan_state.need_index_progress = 1;
 	save_index_scan_state();
 }
-bool get_disk_size(INT64 &total_size,INT64 &free_size) {
+bool get_disk_size(INT64& total_size, INT64& free_size) {
 	KStringBuf path;
 	get_disk_base_dir(path);
 #if defined(_WIN32)
@@ -663,10 +663,10 @@ bool get_disk_size(INT64 &total_size,INT64 &free_size) {
 }
 INT64 get_need_free_disk_size(int used_radio)
 {
-	if (used_radio>=98) {
+	if (used_radio >= 98) {
 		used_radio = 98;
 	}
-	if (used_radio<=0) {
+	if (used_radio <= 0) {
 		used_radio = 1;
 	}
 	INT64 total_size = 0;
